// src/types/onyx/FirebaseSyncStatus.ts
type FirebaseSyncStatus = {
    isUsersSyncCompleted: boolean;
    isRestaurantsSyncCompleted: boolean;
    isEventsSyncCompleted: boolean;
    isRecipesSyncCompleted: boolean;
    isReviewsSyncCompleted: boolean;
    isPhotosSyncCompleted: boolean;
    isPeopleInEventsSyncCompleted: boolean;
};
export default FirebaseSyncStatus;